requirements.txt
flask==3.0.0
PyBluez @ git+https://github.com/pybluez/pybluez.git#egg=pybluez
PyESCPOS[bluetooth]==0.4
pyzmq
pillow
six
